Add support for dotted models
This comes from the need for having fields which access nested parts of
the model. For instance, in VueFormGenerator, you could have a field
with

{
  "schema": {
    "fields": [
      {
        "model": "foo.bar.baz"
      }
    ]
  },
  "model": {
    "foo": {
      "bar": {
        "baz": "value"
      }
    }
  }
}

which would modify the "value" deep in the model object.

Previously, there was no way to write a field that used those nested
models. This commit adds the ability to use double underscores in the
Form to represent a dotted field.

class MyForm(forms.Form):
    foo__bar__baz = forms.CharField()